chaos: move almost all processing into lib.rs
The chaos driver in `main.rs` is getting progressively more complicated as we
add more features, and is only going to get more complicated as causal profiling
joins the fun and we add more modes. Move almost everything into `lib.rs`, leaving
only arg parsing (but not the arg structures) and logger setup in `main.rs`. This
makes it easier to start splitting logic further.

Test plan:
- It builds.
- CI
